refactor(#27): 매물 목록 ViewHolder 이동
